feat：新增語言模組及相關功能
- 在 app.module.ts 中引入 LanguageModule
- 新增 LanguageController，提供語言的 CRUD API 端點
- 新增 LanguageService，實現語言的業務邏輯
- 新增 LanguageEntity，定義語言的資料結構
- 新增請求和回應 DTO，支援語言的資料傳輸
- 更新 DeptReqDto、DictDataReqDto、DictTypeReqDto、MenuReqDto、RoleReqDto、UserReqDto，統一使用 MAX_PAGE_SIZE 和 MAX_PAGE_NUMBER 常數
- 更新 DictTypeEntity 中的排序欄位類型為 float

+++ src/modules/system/dict-type/entity/dict-type.entity.ts
@@ -13,6 +13,6 @@ export class DictTypeEntity extends BaseEntity {
   @Column({ length: 50, comment: '字典類型' })
   type: string
 
-  @Column({ type: 'int', comment: '排序' })
+  @Column({ type: 'float', comment: '排序' })
   sort: number
 }
